pyproject.toml
[tool.poetry]
name = "assistente-juridico"
version = "0.1.0"
description = ""
authors = ["Lucas Oliveira"]
readme = "README.md"
packages = [{include = "assistente_juridico"}]

[tool.poetry.dependencies]
python = "3.12.3"
langchain = "^0.2.5"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
faiss-cpu = "^1.8.0"
altair = "^5.3.0"
tiktoken = "^0.7.0"
langchain-community = "^0.2.5"
openai = "^1.34.0"
langchain-openai = "^0.1.8"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
